+++ lib/nnq/cli/config.rb
@@ -44,6 +44,7 @@ module NNQ
       :parallel,
       :transient,
       :verbose,
+      :timestamps,
       :quiet,
       :echo,
       :scripts,